Modifications to support internal users impersonating other users.
store_user_session() now takes an optional internal_user_id of the
impersonating user.

Since we need to calculate the session expiration timestamp multiple
places in the code, I've moved it to be a function here so that code
isn't duplicative. Also, it now handles the case where you *are*
impersonating a user, since we don't want those sessions to be nearly
as long-lived.